add basic new checkin page
