+++ lib/bonfire_web/controllers/checkin_controller.ex
@@ -1,0 +1,11 @@
+defmodule BonfireWeb.CheckinController do
+  use BonfireWeb, :controller
+
+  def index(conn, _params) do
+    render(conn, "index.html")
+  end
+
+  def new(conn, _params) do
+    render(conn, "new.html")
+  end
+end
