+++ lib/bonfire_web/templates/checkin/new.html.eex
@@ -1,0 +1,7 @@
+<div class="container">
+  <h1 class="title">What are you reading?</h1>
+
+  <div>
+    <%= live_render(@conn, BonfireWeb.Live.NewCheckin, session: %{"user_id" => @current_user.id}) %>
+  </div>
+</div>
