+++ lib/bonfire_web/templates/layout/app.html.eex
@@ -23,11 +23,11 @@
           <%= if Pow.Plug.current_user(@conn) do %>
           <li><a href="/tracks">My books</a></li>
           <li>
-            <%= link to: Routes.reading_state_path(@conn, :new), class: "new-reading-state" do %>
+            <%= link to: Routes.checkin_path(@conn, :new), class: "new-checkin" do %>
               <span class="svg-icon">
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M416 208H272V64c0-17.67-14.33-32-32-32h-32c-17.67 0-32 14.33-32 32v144H32c-17.67 0-32 14.33-32 32v32c0 17.67 14.33 32 32 32h144v144c0 17.67 14.33 32 32 32h32c17.67 0 32-14.33 32-32V304h144c17.67 0 32-14.33 32-32v-32c0-17.67-14.33-32-32-32z"/></svg>
               </span>
-              Add new book
+              Check in
             <% end %>
           </li>
