+++ lib/bonfire_web/templates/reading_state/new.html.eex
@@ -2,6 +2,6 @@
   <h1 class="title">Add new book</h1>
 
   <div>
-    <%= live_render(@conn, BonfireWeb.Live.BookSuggestion, session: %{"user_id" => @current_user.id}) %>
+    <%= live_render(@conn, BonfireWeb.Live.NewBook, session: %{"user_id" => @current_user.id}) %>
   </div>
 </div>
