+++ lib/bonfire_web/views/checkin_view.ex
@@ -1,0 +1,3 @@
+defmodule BonfireWeb.CheckinView do
+  use BonfireWeb, :view
+end
